Put all full text query related stuff in the same place
